refactor: evaluate current project Id depending on user settings and return events for digest page
relates #109
